reafactor and cache serverTranslation
