+++ i18n/index.ts
@@ -1,33 +1,23 @@
 import { createInstance } from "i18next";
 import resourcesToBackend from "i18next-resources-to-backend";
 import { initReactI18next } from "react-i18next/initReactI18next";
-import { getOptions, languages } from "./settings";
-import { headers, cookies } from "next/headers";
-import { pathLanguageDetection } from "./utils";
+import { getOptions } from "./settings";
+import { getCurrentLanguage } from "./utils";
+import { cache } from "react";
+import { logMessage } from "@lib/logger";
 
-export async function serverTranslation(
-  ns?: string | string[],
-  options?: { keyPrefix?: string; lang?: string }
-) {
-  const path = (await headers()).get("x-path") ?? "";
-  const pathLang = pathLanguageDetection(path, languages);
-  const cookieLang = (await cookies()).get("i18next")?.value;
+export const serverTranslation = cache(
+  async (ns?: string | string[], options?: { keyPrefix?: string; lang?: string }) => {
+    const i18nLang = options?.lang ?? (await getCurrentLanguage());
+    logMessage.debug(`Server translation for ${i18nLang} and ns: ${ns ?? "common"}`);
 
-  const i18nLang = options?.lang || pathLang || cookieLang || languages[0];
-
-  const i18nextInstance = await initI18next(i18nLang, ns ?? ["common"]);
-  return {
-    t: i18nextInstance.getFixedT(i18nLang, ns, options?.keyPrefix),
-    i18n: i18nextInstance,
-  };
-}
-
-export async function getCurrentLanguage() {
-  const path = (await headers()).get("x-path") ?? "";
-  const pathLang = pathLanguageDetection(path, languages);
-  const cookieLang = (await cookies()).get("i18next")?.value;
-  return pathLang || cookieLang || languages[0];
-}
+    const i18nextInstance = await initI18next(i18nLang, ns ?? ["common"]);
+    return {
+      t: i18nextInstance.getFixedT(i18nLang, ns, options?.keyPrefix),
+      i18n: i18nextInstance,
+    };
+  }
+);
 
 const initI18next = async (lang: string, ns: string | string[]) => {
   const i18nInstance = createInstance();
