+++ i18n/utils.ts
@@ -1,6 +1,7 @@
 import { languages } from "./settings";
+import { headers, cookies } from "next/headers";
 
-export const pathLanguageDetection = (path: string, languages: string[]) => {
+const pathLanguageDetection = (path: string, languages: string[]) => {
   const pathRegEx = new RegExp(`^/(${languages.join("|")})`);
   const match = path.match(pathRegEx);
   return match ? match[1] : null;
@@ -17,3 +18,10 @@ export const languageParamSanitization = (locale: string | string[] | undefined)
   // If unknown locale, return default locale
   return languages[0];
 };
+
+export async function getCurrentLanguage() {
+  const path = (await headers()).get("x-path") ?? "";
+  const pathLang = pathLanguageDetection(path, languages);
+  const cookieLang = (await cookies()).get("i18next")?.value;
+  return pathLang || cookieLang || languages[0];
+}
